+++ src/ByteDev.Crypto/Encryption/EncryptAttribute.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace ByteDev.Crypto.Encryption
+{
+    /// <summary>
+    /// Indicates that a property should be encrypted.
+    /// </summary>
+    [AttributeUsage(AttributeTargets.Property)]
+    public sealed class EncryptAttribute : Attribute
+    {
+    }
+}