+++ tests/ByteDev.Crypto.UnitTests/Encryption/TestNoAttributes.cs
@@ -1,0 +1,9 @@
+﻿namespace ByteDev.Crypto.UnitTests.Encryption
+{
+    internal class TestNoAttributes
+    {
+        public string Name { get; set; }
+
+        public int Age { get; set; }
+    }
+}